avoided QUnit's global shortcuts
In Node.js, "module" is special, almost like a reserved word; it can
only be redefined within each module individually, but not globally.

Thus tests must not use this shortcut. Adjusting existing test files
should be relatively easy though.

Note that there are alternatives to using an anonymous function wrapper
to create local aliases - e.g. `with(QUnit) { ... }`, but that was
deemed undesirable.